The async tests are back in TextEngine.
As much as I don't like the timer to wait the frames,
I dislike testing against a "testable" sub-class even
more. Only other option to make it more testable would
be to open up the API (bad) or pitch events. TextEngine
IS an EventDispatcher (why?) - seems like some validation
events might be OK. For now I will just test it this way.
It works, it is just kind of burly.